requirements.txt
absl-py==2.1.0
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiortsp==1.4.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==24.2.0
av==15.0.0
backoff==2.2.1
bcrypt==4.3.0
beaupy==3.10.1
beautifulsoup4==4.12.3
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
chromadb==1.0.8
click==8.1.8
coloredlogs==15.0.1
dataclasses-json==0.6.7
datasets==2.21.0
Deprecated==1.2.18
diffusers==0.33.1
dill==0.3.8
distro==1.9.0
dpkt==1.9.8
durationpy==0.9
emoji==2.14.1
fastapi==0.115.9
filelock==3.15.4
flatbuffers==25.2.10
frozenlist==1.4.1
fsspec==2024.6.1
google==3.0.0
google-auth==2.40.1
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.8
ifaddr==0.2.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-text-splitters==0.3.8
langsmith==0.3.42
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.97.0
opencv-python==4.10.0.84
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.18
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
posthog==4.0.1
propcache==0.3.2
protobuf==5.26.1
pupil-labs-neon-recording==1.0.3
pupil-labs-realtime-api==1.6.2
pupil-labs-video==1.0.6
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.9.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-yakh==0.4.1
pytz==2024.1
PyYAML==6.0.2
questo==0.4.1
referencing==0.36.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.4.4
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.45.3
sympy==1.13.2
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==0.4.12
tokenizers==0.21.1
torch==2.4.0
torch-tb-profiler==0.4.3
torchvision==0.19.0
tqdm==4.66.5
transformers==4.51.3
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
universal_pathlib==0.2.6
urllib3==2.2.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zeroconf==0.147.0
zipp==3.21.0
zstandard==0.23.0
